+++ Progetto 2/StrutturaAlberoUtente.h
@@ -6,7 +6,7 @@
 struct Utente {
    char Mail[MAX_STRINGHE];
    char Password[MAX_STRINGHE]; 
-   float Saldo;        
+   int Saldo;        
    struct NodoListaDesideri *ListaDesideri;
    struct Utente *DX; //puntatore al sottoalbero destro
    struct Utente *SX; //puntatore al sottoalbero sinistro
@@ -16,10 +16,10 @@ typedef struct Utente* AlberoUtenti;
 
 //Funzioni gestione Albero Utenti
 AlberoUtenti RicercaMinimoUtenti(AlberoUtenti RadiceInput);
-AlberoUtenti Ins_OrdUtenti(char *MailInput, char *PasswordInput, float SaldoInput,AlberoUtenti RADICE);
+AlberoUtenti Ins_OrdUtenti(char *MailInput, char *PasswordInput, int SaldoInput,AlberoUtenti RADICE);
 void InorderUtenti(AlberoUtenti RADICE);
-AlberoUtenti AggiungiAlSaldo(AlberoUtenti NodoInput, float ValoreInput);
-AlberoUtenti SottraiAlSaldo(AlberoUtenti NodoInput, float ValoreInput);
+AlberoUtenti AggiungiAlSaldo(AlberoUtenti NodoInput, int ValoreInput);
+AlberoUtenti SottraiAlSaldo(AlberoUtenti NodoInput, int ValoreInput);
 AlberoUtenti RicercaBinariaUtenti(char * MailInput, char *PasswordInput,AlberoUtenti RADICE);
 AlberoUtenti LetturaDaFileUtenti(FILE *fp, AlberoUtenti RADICE);
 void CreazioneFILERisultatoUtenti(FILE *fp, AlberoUtenti RADICE);
